feat: 섹션 기준 상태 및 컴포넌트 분리로 리렌더링 최적화
